data type update to handle discord ID's in int form

+++ internal/db/migration/000011_player.sql.up.sql
@@ -1,8 +1,9 @@
 CREATE TABLE IF NOT EXISTS player (
-  id SERIAL PRIMARY KEY, -- this is the discord user id technically...
+  id BIGINT PRIMARY KEY, -- this is the discord user id technically...
   role_id INT references role(id) ON DELETE CASCADE,
   alive BOOLEAN NOT NULL DEFAULT TRUE,
   coins INT NOT NULL DEFAULT 200,
   luck INT NOT NULL DEFAULT 0,
   alignment alignment NOT NULL
 )
+
