+++ internal/db/migration/000012_player_item.sql.up.sql
@@ -1,6 +1,7 @@
 CREATE TABLE IF NOT EXISTS player_item (
-  player_id INT NOT NULL REFERENCES player(id) ON DELETE CASCADE,
+  player_id BIGINT NOT NULL REFERENCES player(id) ON DELETE CASCADE,
   item_id INT NOT NULL,
   quantity INT NOT NULL DEFAULT 1,
   PRIMARY KEY (player_id, item_id)
 );
+
