+++ internal/db/migration/000013_player_status.sql.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS player_status (
-  player_id INT NOT NULL REFERENCES player(id) ON DELETE CASCADE,
+  player_id BIGINT NOT NULL REFERENCES player(id) ON DELETE CASCADE,
   status_id INT NOT NULL,
   quantity INT NOT NULL DEFAULT 1,
   PRIMARY KEY (player_id, status_id)
