+++ internal/db/migration/000014_player_perk.sql.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS player_perk (
-  player_id INT NOT NULL REFERENCES player(id) ON DELETE CASCADE,
+  player_id BIGINT NOT NULL REFERENCES player(id) ON DELETE CASCADE,
   perk_id INT NOT NULL,
   PRIMARY KEY (player_id, perk_id)
 );
