+++ internal/db/migration/000016_player_confessional.sql.up.sql
@@ -1,6 +1,7 @@
 CREATE TABLE IF NOT EXISTS player_confessional (
-  player_id INT NOT NULL REFERENCES player(id) ON DELETE CASCADE,
-  channel_id INT NOT NULL,
-  pin_message_id INT NOT NULL,
+  player_id BIGINT NOT NULL REFERENCES player(id) ON DELETE CASCADE,
+  channel_id BIGINT NOT NULL,
+  pin_message_id BIGINT NOT NULL,
   PRIMARY KEY (player_id, channel_id)
 );
+
